+++ EWARM/NORMA-STM32F103R8T6.ewp
@@ -369,6 +369,7 @@
                     <state>$PROJ_DIR$\..\Src\current_sense</state>
                     <state>$PROJ_DIR$\..\Src\actuators</state>
                     <state>$PROJ_DIR$\..\Src\buttons</state>
+                    <state>$PROJ_DIR$\..\Src\debug</state>
                 </option>
                 <option>
                     <name>CCStdIncCheck</name>
@@ -1118,6 +1119,15 @@
                 </file>
             </group>
             <group>
+                <name>debug</name>
+                <file>
+                    <name>$PROJ_DIR$\..\Src\debug\debug.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\Src\debug\debug.h</name>
+                </file>
+            </group>
+            <group>
                 <name>drivers</name>
                 <file>
                     <name>$PROJ_DIR$\..\Src\drivers\adc.c</name>
@@ -1199,9 +1209,6 @@
         <group>
             <name>User</name>
             <file>
-                <name>$PROJ_DIR$\..\Src\gpio.c</name>
-            </file>
-            <file>
                 <name>$PROJ_DIR$\..\Src\main.c</name>
             </file>
             <file>
@@ -1209,9 +1216,6 @@
             </file>
             <file>
                 <name>$PROJ_DIR$\..\Src\stm32f1xx_it.c</name>
-            </file>
-            <file>
-                <name>$PROJ_DIR$\..\Src\usart.c</name>
             </file>
         </group>
     </group>
