+++ EWARM/NORMA-STM32F103R8T6.ewt
@@ -1233,6 +1233,15 @@
                 </file>
             </group>
             <group>
+                <name>debug</name>
+                <file>
+                    <name>$PROJ_DIR$\..\Src\debug\debug.c</name>
+                </file>
+                <file>
+                    <name>$PROJ_DIR$\..\Src\debug\debug.h</name>
+                </file>
+            </group>
+            <group>
                 <name>drivers</name>
                 <file>
                     <name>$PROJ_DIR$\..\Src\drivers\adc.c</name>
@@ -1314,9 +1323,6 @@
         <group>
             <name>User</name>
             <file>
-                <name>$PROJ_DIR$\..\Src\gpio.c</name>
-            </file>
-            <file>
                 <name>$PROJ_DIR$\..\Src\main.c</name>
             </file>
             <file>
@@ -1324,9 +1330,6 @@
             </file>
             <file>
                 <name>$PROJ_DIR$\..\Src\stm32f1xx_it.c</name>
-            </file>
-            <file>
-                <name>$PROJ_DIR$\..\Src\usart.c</name>
             </file>
         </group>
     </group>
