+++ Src/debug/debug.c
@@ -1,0 +1,53 @@
+/************************************************************************************
+                  ** C **
+Module:
+Author: Unic-Lab <https://unic-lab.ru/>
+************************************************************************************/
+
+#include "debug.h"
+#include "version.h"
+#include "stm32f1xx_hal.h"
+
+//===================================================================================
+
+UART_HandleTypeDef huart1;
+
+//===================================================================================
+
+void debug_init(void)
+{
+	huart1.Instance = USART1;
+	huart1.Init.BaudRate = 115200;
+	huart1.Init.WordLength = UART_WORDLENGTH_8B;
+	huart1.Init.StopBits = UART_STOPBITS_1;
+	huart1.Init.Parity = UART_PARITY_NONE;
+	huart1.Init.Mode = UART_MODE_TX;
+	huart1.Init.HwFlowCtl = UART_HWCONTROL_NONE;
+	huart1.Init.OverSampling = UART_OVERSAMPLING_16;
+	HAL_HalfDuplex_Init(&huart1);
+}
+
+void HAL_UART_MspInit(UART_HandleTypeDef* uartHandle)
+{
+	GPIO_InitTypeDef GPIO_InitStruct = {0};
+
+	if (uartHandle->Instance == USART1)
+	{
+		__HAL_RCC_USART1_CLK_ENABLE();
+		__HAL_RCC_GPIOA_CLK_ENABLE();
+
+		GPIO_InitStruct.Pin = DEBUG_TX_PIN;
+		GPIO_InitStruct.Mode = GPIO_MODE_AF_OD;
+		GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
+		HAL_GPIO_Init(DEBUG_TX_PORT, &GPIO_InitStruct);
+	}
+}
+
+void HAL_UART_MspDeInit(UART_HandleTypeDef* uartHandle)
+{
+	if (uartHandle->Instance == USART1)
+	{
+		__HAL_RCC_USART1_CLK_DISABLE();
+		HAL_GPIO_DeInit(DEBUG_TX_PORT, DEBUG_TX_PIN);
+	}
+}
