+++ Src/debug/debug.h
@@ -4,14 +4,21 @@ Module:
 Author: Unic-Lab <https://unic-lab.ru/>
 ************************************************************************************/
 
-#ifndef _RESET_H_
-#define _RESET_H_
+#ifndef _DEBUG_H_
+#define _DEBUG_H_
 
 //===================================================================================
 
-void reset_iwdg_init(void);
-void reset_iwdg_refresh(void);
+#include "stm32f1xx_hal.h"
 
 //===================================================================================
 
-#endif // _RESET_H_
+extern UART_HandleTypeDef huart1;
+
+//===================================================================================
+
+void debug_init(void);
+
+//===================================================================================
+
+#endif // _DEBUG_H_
