+++ Src/drivers/adc.h
@@ -18,4 +18,6 @@ void adc_ext_volt_deinit(void);
 void adc_current_sense_init(uint32_t* p_buf, uint16_t buf_size);
 void adc_current_senset_deinit(void);
 
+//===================================================================================
+
 #endif // _ADC_H_
