+++ Src/stm32f1xx_it.c
@@ -1,4 +1,3 @@
-#include "main.h"
 #include "stm32f1xx_it.h"
 #include "FreeRTOS.h"
 #include "task.h"
